Alterações semi finais front-end
Arrumei a parte de visitante

+++ crud_connectxpert/app/view/encontro/listEncontro.php
@@ -8,33 +8,32 @@ echo '<link rel="stylesheet" href="' . BASEURL . '/view/turma/turmastyle.css">';
 ?>
 <div class="container">
     <div class="row" style="margin-top: 10px;">
-        <h3 class="col-4">Dados da Turma
-    <a class="btn-inserir" 
-        href="<?= BASEURL ?>/controller/TurmaController.php?action=listProfessor"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-arrow-left-circle" viewBox="0 0 16 16"  style="color: rgba(255, 127, 50, 1);" >
-  <path fill-rule="evenodd" d="M1 8a7 7 0 1 0 14 0A7 7 0 0 0 1 8zm15 0A8 8 0 1 1 0 8a8 8 0 0 1 16 0zm-4.5-.5a.5.5 0 0 1 0 1H5.707l2.147 2.146a.5.5 0 0 1-.708.708l-3-3a.5.5 0 0 1 0-.708l3-3a.5.5 0 1 1 .708.708L5.707 7.5H11.5z"/>
-</svg></a></h3>
+        <h3 class="col-12">Dados da Turma
+            <a class="btn-inserir" href="<?= BASEURL ?>/controller/TurmaController.php?action=listProfessor"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-arrow-left-circle" viewBox="0 0 16 16" style="color: rgba(255, 127, 50, 1);">
+                    <path fill-rule="evenodd" d="M1 8a7 7 0 1 0 14 0A7 7 0 0 0 1 8zm15 0A8 8 0 1 1 0 8a8 8 0 0 1 16 0zm-4.5-.5a.5.5 0 0 1 0 1H5.707l2.147 2.146a.5.5 0 0 1-.708.708l-3-3a.5.5 0 0 1 0-.708l3-3a.5.5 0 1 1 .708.708L5.707 7.5H11.5z" />
+                </svg></a>
+        </h3>
     </div>
     <div class="row" style="margin-top: 10px;">
         <div class="col-12">
-            <span style="font-weight: bold;">Nome: </span> <?= $dados["turma"]->getNomeTurma() ?>
+            <span style="font-weight: bold; color: #ff7f32;">Nome: </span> <?= $dados["turma"]->getNomeTurma() ?>
         </div>
         <div class="col-12">
-            <span style="font-weight: bold;">Horário: </span> <?= $dados["turma"]->getHorario() ?>
+            <span style="font-weight: bold; color: #ff7f32;">Horário: </span> <?= $dados["turma"]->getHorario() ?>
         </div>
     </div>
 </div>
 
 <div class="container">
     <div class="row">
-    <h3 class="col-4">Encontro <a class="btn-inserir" 
-                href="<?= BASEURL ?>/controller/EncontroController.php?action=create&id=<?php echo $dados["idTurma"] ?>">
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-plus-circle" viewBox="0 0 16 16" style="color: rgba(255, 127, 50, 1);">              
-  <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z"/>
-  <path d="M8 4a.5.5 0 0 1 .5.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3A.5.5 0 0 1 8 4z"/>
-  
+        <h3 class="col-12">Encontro <a class="btn-inserir" href="<?= BASEURL ?>/controller/EncontroController.php?action=create&id=<?php echo $dados["idTurma"] ?>">
+                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-plus-circle" viewBox="0 0 16 16" style="color: rgba(255, 127, 50, 1);">
+                    <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z" />
+                    <path d="M8 4a.5.5 0 0 1 .5.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3A.5.5 0 0 1 8 4z" />
 
-  
-</svg></a></h3>
+
+
+                </svg></a></h3>
         <div class="col-9">
             <?php require_once(__DIR__ . "/../include/msg.php"); ?>
         </div>
@@ -42,53 +41,50 @@ echo '<link rel="stylesheet" href="' . BASEURL . '/view/turma/turmastyle.css">';
 
     <div class="row" style="margin-top: 10px;">
         <div class="col-12">
-            <table id="tabEncontro" class='table table-striped'>
-                <thead>
-                    <tr class="topicos">
-                        <th>Nome</th>
-                        <th>Data</th>
-                        <th>Frequência</th>
-                        <th>Alterar</th>
-                        <th>Excluir</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <?php foreach($dados['lista'] as $encontro): ?>
-                        <tr class="listagem">
-                            <td><?= $encontro->getNomeEncontro(); ?></td>
-                            <td><?= $encontro->getDiaEncontroFormatada(); ?></td>
-
-                            <td><a class="btn-alterar" 
-                                href="<?= BASEURL ?>/controller/FrequenciaController.php?action=createFrequencia&id=<?= $encontro->getIdEncontro() ?>&idTurma=<?php echo $dados["idTurma"]; ?>">
-                                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-pencil-square" viewBox="0 0 16 16" style="color: rgba(10, 174, 148, 1);">
-  <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z"/>
-  <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z"/>
-</svg> Frequência </a>                             
-                            <td><a class="btn-alterar" 
-                                href="<?= BASEURL ?>/controller/EncontroController.php?action=edit&id=<?= $encontro->getIdEncontro() ?>">
-                                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-pencil-square" viewBox="0 0 16 16" style="color: rgba(10, 174, 148, 1);">
-  <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z"/>
-  <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z"/>
-</svg> Alterar</a> 
-                            </td>
-
-                            <td><a class="btn-excluir" 
-                                onclick="return confirm('Confirma a exclusão do encontro? Ao excluir este encontro todos as frequências atreladas a eles serão excluídas');"
-                                href="<?= BASEURL ?>/controller/EncontroController.php?action=delete&id=<?= $encontro->getIdEncontro() ?>">
-                                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
-  <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z"/>
-  <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z"/>
-</svg> Excluir</a> 
-                            </td>
+            <div class="table-responsive">
+                <table id="tabEncontro" class='table table-striped'>
+                    <thead>
+                        <tr class="topicos">
+                            <th>Nome</th>
+                            <th>Data</th>
+                            <th>Frequência</th>
+                            <th>Alterar</th>
+                            <th>Excluir</th>
                         </tr>
-                    <?php endforeach; ?>
-                </tbody>
-            </table>
+                    </thead>
+                    <tbody>
+                        <?php foreach ($dados['lista'] as $encontro) : ?>
+                            <tr class="listagem">
+                                <td><?= $encontro->getNomeEncontro(); ?></td>
+                                <td><?= $encontro->getDiaEncontroFormatada(); ?></td>
+
+                                <td><a class="btn-alterar" href="<?= BASEURL ?>/controller/FrequenciaController.php?action=createFrequencia&id=<?= $encontro->getIdEncontro() ?>&idTurma=<?php echo $dados["idTurma"]; ?>">
+                                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-pencil-square" viewBox="0 0 16 16" style="color: rgba(10, 174, 148, 1);">
+                                            <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
+                                            <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z" />
+                                        </svg> Frequência </a>
+                                <td><a class="btn-alterar" href="<?= BASEURL ?>/controller/EncontroController.php?action=edit&id=<?= $encontro->getIdEncontro() ?>">
+                                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-pencil-square" viewBox="0 0 16 16" style="color: rgba(10, 174, 148, 1);">
+                                            <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
+                                            <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z" />
+                                        </svg> Alterar</a>
+                                </td>
+
+                                <td><a class="btn-excluir" onclick="return confirm('Confirma a exclusão do encontro? Ao excluir este encontro todos as frequências atreladas a eles serão excluídas');" href="<?= BASEURL ?>/controller/EncontroController.php?action=delete&id=<?= $encontro->getIdEncontro() ?>">
+                                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
+                                            <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z" />
+                                            <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z" />
+                                        </svg> Excluir</a>
+                                </td>
+                            </tr>
+                        <?php endforeach; ?>
+                    </tbody>
+                </table>
+            </div>
         </div>
     </div>
 </div>
 
-<?php  
+<?php
 require_once(__DIR__ . "/../include/footer.php");
 ?>
-
